backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
Flask==3.1.1
python-dotenv==1.0.0
Flask-Login==0.6.3
flask-cors==6.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
transformers==4.36.0
requests==2.31.0
ollama==0.1.7
nltk==3.8.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.9.0
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.3
sendgrid==6.11.0
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
stripe==10.12.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
Werkzeug==3.1.3
wsproto==1.2.0
